chore: update deploy
