+++ .github/workflows/deploy-app.yml
@@ -6,18 +6,26 @@ on:
       env:
         required: true
         type: string
-
-    secrets:
       app_id_github:
         required: true
+        type: string
       pr_author:
         required: true
+        type: string
       project_number:
         required: true
+        type: string
       column_name:
         required: true
+        type: string
+      webhook_path:
+        required: true
+        type: string
       project_id:
         required: true
+        type: string
+
+    secrets:
       gcp_provider_name:
         required: true
       gcp_sa:
@@ -26,14 +34,13 @@ on:
         required: true
       gcp_app_key:
         required: true
-      
 
 jobs:
     deploy:
         name: Setup Gcloud Account
         permissions:
           id-token: write
-          contents: read 
+          contents: read
         runs-on: ubuntu-latest
         env:
           gcp_service_region: us-central1
@@ -52,31 +59,29 @@ jobs:
             workload_identity_provider: ${{ secrets.gcp_provider_name }}
             service_account: ${{ secrets.gcp_sa }}
 
-        - name: Update the GPC Webhook Secrets
-          run: |
-            echo -n "${{ secrets.gcp_webhook_secret }}" | gcloud secrets versions add unbreak_${{ inputs.env }}_webhook_secret --data-file=-
-                       
-            
-        - name: Update the GPC App Key Secrets
-          run: |
-            echo "${{ secrets.gcp_app_key }}" | gcloud secrets versions add unbreak_${{ inputs.env }}_app_key --data-file=-
-
-        - name: Deploy Docker image
-          uses: 'google-github-actions/deploy-cloudrun@v1'
+        - name: Update secrets in GCP Secret Manager
+          uses: nearform-actions/github-action-gcp-secrets@v1
           with:
-            service: 'unbreakci-${{ inputs.env }}'
-            region: ${{ env.gcp_service_region }}
-            project_id: ${{ env.project_id }}
-            flags: --allow-unauthenticated
-            suffix: ${{ github.sha }}
-            source: .
-            env_vars: |
-              APP_ID_GITHUB=${{ secrets.app_id_github }}
-              PR_AUTHOR=${{ secrets.pr_author }}
-              PROJECT_NUMBER=${{ secrets.project_number }}
-              COLUMN_NAME=${{ secrets.column_name }}
-              LOG_LEVEL=info
-            secrets: |
-              WEBHOOK_SECRET=unbreak_${{ inputs.env }}_webhook_secret:latest
-              APP_KEY=unbreak_${{ inputs.env }}_app_key:latest
+            secrets: |-
+              unbreak_${{ inputs.env }}_webhook_secret:"${{ secrets.gcp_webhook_secret }}"
+              unbreak_${{ inputs.env }}_app_key:"${{ secrets.gcp_app_key }}"
 
+        - name: 'Deploy webhook to Cloud Run'
+          id: deploy-webhook
+          run: >-
+            gcloud functions deploy ${{ inputs.webhook_path }}
+            --project="${{ inputs.gcp_project_id }}"
+            --runtime=nodejs18
+            --source=.
+            --entry-point=webhook
+            --set-secrets=APP_KEY=projects/${{ inputs.gcp_project_id }}/secrets/unbreak_${{ inputs.env }}_app_key:latest
+            --set-secrets=WEBHOOK_SECRET=projects/${{ inputs.gcp_project_id }}/secrets/unbreak_${{ inputs.env }}_webhook_secret:latest
+            --set-env-vars=APP_ID="${{ inputs.APP_ID_GITHUB  }}"
+            --set-env-vars=PR_AUTHOR="${{ inputs.pr_author }}"
+            --set-env-vars=PROJECT_NUMBER="${{ inputs.project_number }}"
+            --set-env-vars=COLUMN_NAME="${{ inputs.column_name }}"
+            --set-env-vars=LOG_LEVEL=info
+            --trigger-http
+            --allow-unauthenticated
+            --region=us-central1
+            --gen2
