+++ package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "dev": "nodemon index.js",
-    "start": "node index.js",
+    "dev": "nodemon indexLocal.js",
+    "start": "node indexLocal.js",
     "lint": "eslint .",
     "test": "jest",
     "prepare": "npx husky install"
